medical_agent/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
streamlit==1.32.2
langchain==0.3.1
langchain-community==0.3.1
langgraph==0.2.20
langchain-ollama==0.2.0
chromadb>=0.6.0
neo4j==5.24.0
pydantic==2.9.2
pymysql==1.1.0
cryptography==42.0.5
pandas>=2.0.0
plotly>=5.18.0
networkx>=3.1
pyvis>=0.3.2
pdfplumber>=0.11.0
python-docx>=1.1.0
openpyxl>=3.1.2
python-dotenv==1.0.1
requests==2.32.3
streamlit-option-menu
tenacity>=8.2.0
DBUtils>=3.1.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
